chore: rename context module to avoid import collision
